requirements.txt
requests
python-dotenv
google-generativeai
openai
anthropic
pydantic
mcp
pytest
